+++ docs/conf.py
@@ -59,7 +59,7 @@ copyright = '2012 - 2016, Michal Čihař'
 # built documents.
 #
 # The short X.Y version.
-version = '2.6'
+version = '2.7'
 # The full version, including alpha/beta/rc tags.
 release = version