+++ docs/changes.rst
@@ -4,7 +4,7 @@ Changes
 weblate 2.6
 -----------
 
-Released on ? 2016.
+Released on Apr 28th 2016.
 
 * Fixed validation of subprojects with language filter.
 * Improved support for XLIFF files.